refactor: Update deleteRoom action to disassociate users from the room and delete the room
